requirements.txt.txt
openai
python-dotenv
pdfplumber